+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Added `run` command to cli for performing both data collection and
+  synchronization.
+- Expose `createApiClient` and synchronization functions
+
 ## 0.13.0 2020-04-29
 
 ### Added